pyproject.toml
[project]
name = "fridge"
version = "0.1.0"
description = "Noise detector using audio analysis"
requires-python = ">=3.12"
dependencies = [
  "numpy>=1.26",
  "pyyaml>=6.0",
  "scipy>=1.11",
  "scikit-learn>=1.4",
  "soundfile>=0.12",
  "sounddevice>=0.4.6",
  "torch>=2.2",
  "torchaudio>=2.2",
  "tqdm>=4.66",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
